Move tests out into tests/
- Hard code some assumptions
